Added authorization on pdate and destroy methods.

+++ routes/api.php
@@ -27,12 +27,12 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 /* ***************************************************************** Admin routes ***************************************************************** */
 Route::get("/v1/users", [AuthController::class, "index"]);
-Route::post("/v1/register/admin", [AuthController::class, 'registerAdmin'])->middleware("auth:sanctum");;
+Route::post("/v1/register/admin", [AuthController::class, 'registerAdmin'])->middleware("auth:sanctum");
 Route::post("/v1/register/user", [AuthController::class, 'registerUser']);
 Route::post("/v1/login", [AuthController::class, 'login']);
 Route::post("/v1/user_info", [AuthController::class, 'userInfo'])->middleware("auth:sanctum");
-Route::put("/v1/user/{user}", [AuthController::class, "update"]);
-Route::delete("/v1/user/{user}", [AuthController::class, "destroy"]);
+Route::put("/v1/user/{user}", [AuthController::class, "update"])->middleware("auth:sanctum");
+Route::delete("/v1/user/{user}", [AuthController::class, "destroy"])->middleware("auth:sanctum");
 /* ************************************************************************************************************************************************ */
 
 /* **************************************************************** Recipes routes **************************************************************** */
